+++ src/donnees/Donnees.java
@@ -1,4 +1,4 @@
-package partie2.donnees;
+package donnees;
 
 import java.util.ArrayList;
