+++ src/mlp/Sigmoid.java
@@ -1,6 +1,6 @@
-package partie1;
+package mlp;
 
-class Sigmoid implements TransferFunction {
+public class Sigmoid implements TransferFunction {
     @Override
     public double evaluate(double value) {
         return 1.0 / (1.0 + Math.exp(-value));
